+++ hyoga/open/reprojected.py
@@ -74,8 +74,7 @@ def _open_climatology(source='chelsa', variable='tas'):
     elif source == 'cw5e5':
         aggregator = hyoga.open.aggregator.CW5E5TiledAggregator()
         start, end = 1981, 2010  # FIXME allow custom aggregation period
-        paths = (aggregator(variable, start, end, mon) for mon in range(1, 13))
-        paths = [path for month in paths for path in month]
+        paths = aggregator(variable, start, end)
         ds = xr.open_mfdataset(paths, decode_cf=True)
         da = ds[variable].rio.write_crs('+proj=longlat +datum=WGS84')