+++ hyoga/open/reprojected.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2022-2024, Julien Seguinot (juseg.dev)
+# Copyright (c) 2022-2025, Julien Seguinot (juseg.dev)
 # GNU General Public License v3.0+ (https://www.gnu.org/licenses/gpl-3.0.txt)
 
 """