+++ doc/whatsnew.rst
@@ -32,7 +32,8 @@ New features
 - Add ``temperature`` and ``precipitation`` arguments, and CHELSA-W5E5_
   (aliased ``'cw5e5'``) as a new temperature and precipitation data source in
   :func:`hyoga.open.atmosphere` (:issue:`86`, :pull:`87`), aggregated on 30x30
-  degree tiles (:issue:`88`, :pull:`92`).
+  degree tiles (:issue:`88`, :pull:`92`), concatenated yearly (:issue:`99`,
+  :pull:`100`).
 
 .. _CHELSA-W5E5: https://chelsa-climate.org/chelsa-w5e5-v1-0-daily-climate-data-at-1km-resolution/
 
@@ -45,7 +46,7 @@ Internal changes
 ~~~~~~~~~~~~~~~~
 
 - Add aggregators in :mod:`hyoga.open.aggregator` (:issue:`86`, :issue:`88`,
-  :pull:`87`, :pull:`92`).
+  :issue:`99`, :pull:`87`, :pull:`92`, :pull:`100`).
 
 .. _v0.3.1: