+++ doc/whatsnew.rst
@@ -42,6 +42,7 @@ Bug fixes
 
 - Add newly missing requirement of ``numpy<2`` (:issue:`90`, :pull:`91`).
 - Fix CHELSA_-2.1 download urls (:issue:`103`, :pull:`104`).
+- Fix Ehlers et al. (2011) download error (:issue:`109`, :pull:`110`).
 
 Internal changes
 ~~~~~~~~~~~~~~~~