+++ doc/whatsnew.rst
@@ -43,6 +43,7 @@ Bug fixes
 - Add newly missing requirement of ``numpy<2`` (:issue:`90`, :pull:`91`).
 - Fix CHELSA_-2.1 download urls (:issue:`103`, :pull:`104`).
 - Fix Ehlers et al. (2011) download error (:issue:`109`, :pull:`110`).
+- Add Sphinx config path in `.readthedocs.yaml` (:issue:`105`, :pull:`106`).
 
 Internal changes
 ~~~~~~~~~~~~~~~~